requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
requests~=2.24.0
Appium-Python-Client~=1.0.1
selenium~=3.141.0
pexpect~=4.8.0